refactor: Apply rector Nextcloud 27 set
