+++ tests/lib/AppFramework/Middleware/Security/RateLimitingMiddlewareTest.php
@@ -14,9 +14,6 @@ use OC\AppFramework\Utility\ControllerMethodReflector;
 use OC\Security\Ip\BruteforceAllowList;
 use OC\Security\RateLimiting\Exception\RateLimitExceededException;
 use OC\Security\RateLimiting\Limiter;
-use OCP\AppFramework\Controller;
-use OCP\AppFramework\Http\Attribute\AnonRateLimit;
-use OCP\AppFramework\Http\Attribute\UserRateLimit;
 use OCP\AppFramework\Http\DataResponse;
 use OCP\AppFramework\Http\TemplateResponse;
 use OCP\IAppConfig;
@@ -25,34 +22,8 @@ use OCP\ISession;
 use OCP\IUser;
 use OCP\IUserSession;
 use PHPUnit\Framework\MockObject\MockObject;
+use Test\AppFramework\Middleware\Security\Mock\RateLimitingMiddlewareController;
 use Test\TestCase;
-
-class TestRateLimitController extends Controller {
-	/**
-	 * @UserRateThrottle(limit=20, period=200)
-	 * @AnonRateThrottle(limit=10, period=100)
-	 */
-	public function testMethodWithAnnotation() {
-	}
-
-	/**
-	 * @AnonRateThrottle(limit=10, period=100)
-	 */
-	public function testMethodWithAnnotationFallback() {
-	}
-
-	public function testMethodWithoutAnnotation() {
-	}
-
-	#[UserRateLimit(limit: 20, period: 200)]
-	#[AnonRateLimit(limit: 10, period: 100)]
-	public function testMethodWithAttributes() {
-	}
-
-	#[AnonRateLimit(limit: 10, period: 100)]
-	public function testMethodWithAttributesFallback() {
-	}
-}
 
 /**
  * @group DB
@@ -101,8 +72,8 @@ class RateLimitingMiddlewareTest extends TestCase {
 			->method('isLoggedIn')
 			->willReturn(false);
 
-		/** @var TestRateLimitController|MockObject $controller */
-		$controller = $this->createMock(TestRateLimitController::class);
+		/** @var RateLimitingMiddlewareController|MockObject $controller */
+		$controller = $this->createMock(RateLimitingMiddlewareController::class);
 		$this->reflector->reflect($controller, 'testMethodWithoutAnnotation');
 		$this->rateLimitingMiddleware->beforeController($controller, 'testMethodWithoutAnnotation');
 	}
@@ -119,14 +90,14 @@ class RateLimitingMiddlewareTest extends TestCase {
 			->method('isLoggedIn')
 			->willReturn(true);
 
-		/** @var TestRateLimitController|MockObject $controller */
-		$controller = $this->createMock(TestRateLimitController::class);
+		/** @var RateLimitingMiddlewareController|MockObject $controller */
+		$controller = $this->createMock(RateLimitingMiddlewareController::class);
 		$this->reflector->reflect($controller, 'testMethodWithoutAnnotation');
 		$this->rateLimitingMiddleware->beforeController($controller, 'testMethodWithoutAnnotation');
 	}
 
 	public function testBeforeControllerForAnon(): void {
-		$controller = new TestRateLimitController('test', $this->request);
+		$controller = new RateLimitingMiddlewareController('test', $this->request);
 
 		$this->request
 			->method('getRemoteAddress')
@@ -150,7 +121,7 @@ class RateLimitingMiddlewareTest extends TestCase {
 	}
 
 	public function testBeforeControllerForLoggedIn(): void {
-		$controller = new TestRateLimitController('test', $this->request);
+		$controller = new RateLimitingMiddlewareController('test', $this->request);
 		/** @var IUser|MockObject $user */
 		$user = $this->createMock(IUser::class);
 
@@ -177,7 +148,7 @@ class RateLimitingMiddlewareTest extends TestCase {
 	}
 
 	public function testBeforeControllerAnonWithFallback(): void {
-		$controller = new TestRateLimitController('test', $this->request);
+		$controller = new RateLimitingMiddlewareController('test', $this->request);
 		$this->request
 			->expects($this->once())
 			->method('getRemoteAddress')
@@ -202,7 +173,7 @@ class RateLimitingMiddlewareTest extends TestCase {
 	}
 
 	public function testBeforeControllerAttributesForAnon(): void {
-		$controller = new TestRateLimitController('test', $this->request);
+		$controller = new RateLimitingMiddlewareController('test', $this->request);
 
 		$this->request
 			->method('getRemoteAddress')
@@ -226,7 +197,7 @@ class RateLimitingMiddlewareTest extends TestCase {
 	}
 
 	public function testBeforeControllerAttributesForLoggedIn(): void {
-		$controller = new TestRateLimitController('test', $this->request);
+		$controller = new RateLimitingMiddlewareController('test', $this->request);
 		/** @var IUser|MockObject $user */
 		$user = $this->createMock(IUser::class);
 
@@ -253,7 +224,7 @@ class RateLimitingMiddlewareTest extends TestCase {
 	}
 
 	public function testBeforeControllerAttributesAnonWithFallback(): void {
-		$controller = new TestRateLimitController('test', $this->request);
+		$controller = new RateLimitingMiddlewareController('test', $this->request);
 		$this->request
 			->expects($this->once())
 			->method('getRemoteAddress')
@@ -281,13 +252,13 @@ class RateLimitingMiddlewareTest extends TestCase {
 		$this->expectException(\Exception::class);
 		$this->expectExceptionMessage('My test exception');
 
-		$controller = new TestRateLimitController('test', $this->request);
+		$controller = new RateLimitingMiddlewareController('test', $this->request);
 
 		$this->rateLimitingMiddleware->afterException($controller, 'testMethod', new \Exception('My test exception'));
 	}
 
 	public function testAfterExceptionWithJsonBody(): void {
-		$controller = new TestRateLimitController('test', $this->request);
+		$controller = new RateLimitingMiddlewareController('test', $this->request);
 		$this->request
 			->expects($this->once())
 			->method('getHeader')
@@ -301,7 +272,7 @@ class RateLimitingMiddlewareTest extends TestCase {
 	}
 
 	public function testAfterExceptionWithHtmlBody(): void {
-		$controller = new TestRateLimitController('test', $this->request);
+		$controller = new RateLimitingMiddlewareController('test', $this->request);
 		$this->request
 			->expects($this->once())
 			->method('getHeader')
